requirements.txt
appdirs==1.4.4
asgiref==3.3.1
black==20.8b1
click==7.1.2
Django==3.1.7
django-extensions==3.1.1
djangorestframework==3.12.4
et-xmlfile==1.0.1
flake8==3.9.0
isort==5.8.0
Markdown==3.3.4
mccabe==0.6.1
mypy-extensions==0.4.3
openpyxl==3.0.7
pathspec==0.8.1
psycopg2-binary==2.8.6
pycodestyle==2.7.0
pyflakes==2.3.1
python-dotenv==0.17.0
pytz==2021.1
regex==2021.3.17
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
